:bug: [User] Move email from baseuser to user
